Models refactoring
