+++ GoXLR TouchPortal Plugin.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30711.63
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoXLR.Worker", "GoXLR TouchPortal Plugin\GoXLR.Worker.csproj", "{05324A5A-A068-4350-9C2D-454D5848EAE4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoXLR.Worker", "GoXLR TouchPortal Plugin\GoXLR.Worker.csproj", "{05324A5A-A068-4350-9C2D-454D5848EAE4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoXLR.Models", "GoXLR.Models\GoXLR.Models.csproj", "{F51ECBBC-8078-4D85-A0C4-1DA2537B6532}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{05324A5A-A068-4350-9C2D-454D5848EAE4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{05324A5A-A068-4350-9C2D-454D5848EAE4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{05324A5A-A068-4350-9C2D-454D5848EAE4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F51ECBBC-8078-4D85-A0C4-1DA2537B6532}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F51ECBBC-8078-4D85-A0C4-1DA2537B6532}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F51ECBBC-8078-4D85-A0C4-1DA2537B6532}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F51ECBBC-8078-4D85-A0C4-1DA2537B6532}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
