+++ GoXLR.Models/Models/Payloads/GetProfilesResponsePayload.cs
@@ -1,0 +1,7 @@
+﻿namespace GoXLR.Models.Models.Payloads
+{
+    public class GetProfilesResponsePayload
+    {
+        public string[] Profiles { get; set; }
+    }
+}
