+++ GoXLR.Models/Models/Shared/ModelBase.cs
@@ -1,0 +1,16 @@
+﻿using System.Text.Json.Serialization;
+
+namespace GoXLR.Models.Models.Shared
+{
+    public class ModelBase
+    {
+        [JsonPropertyName("action")]
+        public string Action { get; set; }
+
+        [JsonPropertyName("context")]
+        public string Context { get; set; }
+
+        [JsonPropertyName("event")]
+        public string Event { get; set; }
+    }
+}
