+++ GoXLR.Models/Models/Shared/RequestModelBase.cs
@@ -1,0 +1,11 @@
+﻿using System.Text.Json.Serialization;
+
+namespace GoXLR.Models.Models.Shared
+{
+    public class RequestModelBase : ModelBase
+    {
+        //The id of the controller device?
+        [JsonPropertyName("device")]
+        public string Device { get; set; }
+    }
+}
